pyproject.toml
[project]
name = "sarcastic-bot"
version = "0.1.0"
description = "Саркастический Telegram-бот для консультаций"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.7.0",
    "aiohttp>=3.9.5",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "black>=24.4.2",
    "flake8>=7.0.0",
    "mypy>=1.10.0",
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
